LayoutTests/platform/ios/ios/touch/touch-handler-on-window-expected.txt

CONSOLE MESSAGE: line 22: TypeError: eventSender.clearTouchPoints is not a function. (In 'eventSender.clearTouchPoints()', 'eventSender.clearTouchPoints' is undefined)
Tests that touch events are dispatched to the DOMWindow even if no other elements have touch event listeners.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
